requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
redis>=5.2.0
httpx>=0.27.0
pydantic>=2.9.0
python-telegram-bot>=21.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17
aiohttp>=3.11.0
websockets>=13.0
flask>=3.0.0
flask-cors>=5.0.0
gunicorn>=23.0.0
requests>=2.32.0
numpy>=1.26.0
pandas>=2.2.0
cloudscraper>=1.2.71
eth-account>=0.13.0
pynacl>=1.5.0
base58>=2.1.0
